feat: finish langchain rework
